Checkout | Move views into view folder
