+++ packages/checkout/widgets/src/widgets/connect/views/ConnectResult.tsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
-import { FooterLogo } from '../../../../components/Footer/FooterLogo';
-import { HeaderNavigation } from '../../../../components/Header/HeaderNavigation';
-import { SimpleLayout } from '../../../../components/SimpleLayout/SimpleLayout';
-import { ViewContext } from '../../../../context/ViewContext';
-import { ConnectWidgetViews } from '../../../../context/ConnectViewContextTypes';
+import { FooterLogo } from '../../../components/Footer/FooterLogo';
+import { HeaderNavigation } from '../../../components/Header/HeaderNavigation';
+import { SimpleLayout } from '../../../components/SimpleLayout/SimpleLayout';
+import { ViewContext } from '../../../context/ViewContext';
+import { ConnectWidgetViews } from '../../../context/ConnectViewContextTypes';
 import { Body } from '@biom3/react';
 
 export const ConnectResult = () => {
